requirements.txt
torch>=2.0.0
scipy>=1.10.0
numpy>=1.24.0
matplotlib>=3.7.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tqdm>=4.60.0
